wxGUI: fix show MASK statusbar button widget if mask is created (#2520)
Add standalone watchdogs into single/multi window mode which handle
changes in the current mapset and current mapset maps and show/hide MASK
statusbar button widget according this state.
